+++ biblioteca/src/main/java/com/residencia/biblioteca/services/LivroService.java
@@ -10,26 +10,43 @@ import com.residencia.biblioteca.repositories.LivroRepository;
 
 @Service
 public class LivroService {
-	@Autowired  
+	@Autowired
 	LivroRepository LivroRepo;
-	
-	public List<Livro> listarLivros(){
+
+	public List<Livro> listarLivros() {
 		return LivroRepo.findAll();
 	}
-	
+
 	public Livro buscarLivroPorId(Integer id) {
 		return LivroRepo.findById(id).orElse(null);
 	}
-	
+
 	public Livro salvarLivro(Livro livro) {
 		return LivroRepo.save(livro);
 	}
-	
+
 	public Livro atualizarLivro(Livro livro) {
 		return LivroRepo.save(livro);
 	}
-	
-	public void deletarLivro(Livro livro) {
-		 LivroRepo.delete(livro);
+
+	public Boolean deletarLivro(Livro livro) {
+		if (livro == null) {
+			return false;
+		}
+		Livro livroExistente = buscarLivroPorId(livro.getCodigoLivro());
+
+		if (livroExistente == null) {
+			return false;
+		}
+
+		LivroRepo.delete(livro);
+
+		Livro livroContinuaExistindo = buscarLivroPorId(livro.getCodigoLivro());
+
+		if (livroContinuaExistindo == null) {
+			return true;
+		}
+
+		return false;
 	}
 }
